requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
slowapi==0.1.9
pydantic==2.5.0
